renaming to menuImages
