+++ backend/app.js
@@ -42,7 +42,7 @@ app.use("/item", require("./routes/item"));
 app.use("/jwt", require("./routes/jwt"));
 app.use("/cart", require("./routes/cart"));
 app.use("/order", require("./routes/order"));
-app.use("/menuItems", require("./routes/menuItems"));
+app.use("/menuImages", require("./routes/menuImages"));
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
