+++ backend/db/models/menuImages.js
@@ -1,0 +1,8 @@
+const mongoose = require("mongoose");
+
+const menuImageSchema = new mongoose.Schema({
+  imageUrl: { type: String, required: true },
+});
+
+const MenuImage = mongoose.model("menuImage", menuImageSchema);
+module.exports = { MenuImage };
