--- backend/db/models/menuItems.js
@@ -1,8 +1,0 @@
-const mongoose = require("mongoose");
-
-const menuItemSchema = new mongoose.Schema({
-  imageUrl: { type: String, required: true },
-});
-
-const MenuItem = mongoose.model("MenuItem", menuItemSchema);
-module.exports = { MenuItem };
