+++ backend/routes/menuImages.js
@@ -2,9 +2,12 @@ const express = require("express");
 const { body } = require("express-validator");
 const { isValidated } = require("../middleware/validation");
 const router = express.Router();
-const { changeMenuImage, findMenuImage } = require("../db/services/menuItems");
+const { changeMenuImage, findMenuImage } = require("../db/services/menuImages");
 
-// @body: imageUrl, returns success:true if imageUrl is changed
+// @body: imageUrl
+// @return: success:true if imageUrl is changed
+//          "MenuImage change unsuccessful" if duplicate imageUrl
+// @description: changes the menu image in the DB
 router.post(
   "/changeMenuImage",
   [body("imageUrl").notEmpty(), isValidated],
@@ -30,7 +33,8 @@ router.post(
   }
 );
 
-// returns the imageUrl in the DB
+// @return: imageUrl object
+// @description: returns the imageUrl in the DB
 router.get("/", async (req, res, next) => {
   try {
     // returns image/null or error if there is an error
